requirements.txt
flask=2.2.2
Werkzeug=2.2.2
pymongo=4.3.2
python-dotenv=0.20.0
pandas=1.4.3
bson=0.5.10
pyodbc=4.0.34